Distinguish MatchParen and Visual.
There are circumstances in which a user may want to visually select text
that is highlighted using MatchParen. Before this commit, MatchParen and
Visual used the same color in dark color schemes. Fix that by
differentiating the colors of the two highlight groups.

Fixes https://github.com/lifepillar/vim-gruvbox8/issues/31.